show alternative code organization around provider and extractor
